K1 - a Wordpress theme
K1 is a free, bespoke, simple theme for Wordpress. It is responsive and
blog-focused, with two widget areas on the main page as well as three in
the footer. The top banner image, text and link can be customised via
the theme customizer, as can the front page ‘About Our Work’ section.